feature/movement-with-collision

+++ Multiplayer-Roguelike.Unity/Assets/Runtime/Ecs/Core/QueryBuffer.cs
@@ -91,9 +91,9 @@ namespace Runtime.Ecs.Core
         where T4 : IComponent
         where T5 : IComponent
     {
-        public int Count = 0;
+        public ushort Count = 0;
 
-        public int[] EntityIds;
+        public ushort[] EntityIds;
 
         public T1[] Components1;
         public T2[] Components2;
@@ -101,9 +101,9 @@ namespace Runtime.Ecs.Core
         public T4[] Components4;
         public T5[] Components5;
 
-        public QueryBuffer(int initialCapacity = 32)
+        public QueryBuffer(ushort initialCapacity = 32)
         {
-            EntityIds = new int[initialCapacity];
+            EntityIds = new ushort[initialCapacity];
             Components1 = new T1[initialCapacity];
             Components2 = new T2[initialCapacity];
             Components3 = new T3[initialCapacity];
